fix: pokemon

+++ app/Http/Controllers/PokemonController.php
@@ -3,8 +3,13 @@
 namespace App\Http\Controllers;
 
 use App\Http\Resources\PokemonResource;
+use App\Models\EncounterAreas;
 use App\Models\Pokemon;
+use App\Models\PokemonAbility;
+use App\Models\PokemonType;
+use App\Models\Species;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
 
 class PokemonController extends Controller
 {
@@ -34,8 +39,8 @@ class PokemonController extends Controller
     public function store(Request $request)
     {
         $validated = $request->validate([
-            'pokemon_id' => ['required', 'integer'],
-            'name' => ['required'],
+            'pokemon_id' => ['required', 'integer', 'unique:pokemon'],
+            'name' => ['required', 'unique:pokemon'],
             'sprite_front' => ['required'],
             'sprite_back' => ['required'],
             'artwork' => ['required'],
@@ -45,6 +50,8 @@ class PokemonController extends Controller
             'stat_special_attack' => ['required', 'integer'],
             'stat_special_defense' => ['required', 'integer'],
             'stat_speed' => ['required', 'integer'],
+            'height' => ['required', 'integer'],
+            'weight' => ['required', 'integer'],
         ]);
 
         return new PokemonResource(Pokemon::create($validated));
@@ -53,8 +60,8 @@ class PokemonController extends Controller
     public function update(Request $request, Pokemon $pokemon)
     {
         $validated = $request->validate([
-            'pokemon_id' => ['required', 'integer'],
-            'name' => ['required'],
+            'pokemon_id' => ['required', 'integer', 'unique:pokemon'],
+            'name' => ['required', 'unique:pokemon'],
             'sprite_front' => ['required'],
             'sprite_back' => ['required'],
             'artwork' => ['required'],
@@ -64,7 +71,18 @@ class PokemonController extends Controller
             'stat_special_attack' => ['required', 'integer'],
             'stat_special_defense' => ['required', 'integer'],
             'stat_speed' => ['required', 'integer'],
+            'height' => ['required', 'integer'],
+            'weight' => ['required', 'integer'],
         ]);
+
+        $isInUse = false;
+        if (PokemonType::where('pokemon', $pokemon->name)->exists() || PokemonAbility::where('pokemon', $pokemon->name)->exists() || EncounterAreas::where('pokemon', $pokemon->name)->exists() || Species::where('pokemon_name', $pokemon->name)->exists()){
+            $isInUse = true;
+        }
+
+        if ($isInUse) {
+            return errorJson('Pokemon is in use and can\'t be updated', 400);
+        }
 
         $pokemon->update($validated);
 
@@ -73,6 +91,13 @@ class PokemonController extends Controller
 
     public function destroy(Pokemon $pokemon)
     {
+        $isInUse = false;
+        if (PokemonType::where('pokemon', $pokemon->name)->exists() || PokemonAbility::where('pokemon', $pokemon->name)->exists() || EncounterAreas::where('pokemon', $pokemon->name)->exists() || Species::where('pokemon_name', $pokemon->name)->exists()){
+            $isInUse = true;
+        }
+        if ($isInUse) {
+            return errorJson('Pokemon is in use and can\'t be deleted', 400);
+        }
         $pokemon->delete();
 
         return response()->json();
